Fixed /ms set limit none and made db_plain care less about convert exceptions

+++ include/services.h
@@ -485,7 +485,7 @@ class Memo : public Flags<MemoFlag>
 
 struct CoreExport MemoInfo
 {
-	unsigned memomax;
+	int16 memomax;
 	std::vector<Memo *> memos;
 
 	unsigned GetIndex(Memo *m) const;
